Fixed a bug with paths
Fixed a bug where using the plugin had to be done in the plugin's
working directory. This meant that unless you have the 'plugin' folder
in your root, the plugin won't work because of stupid paths issues, so
I've finally fixed this.

Another fix was changing the 'g:' scope of variables to 's:'.